Move VariableNotDeclaredIsnpectionTests for QuickFixes into relevant files under RD.QuickFixes
